+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Version 0.1.35
 -------------
 - #374 Support pyzbar
 - #372 Support pydmtxlib
+- #368 Ugly error message when creating a new inselect document on Mac OS X
 - #273 Inselect as a package
 - #83 Architecture and code organisation